[Windows] Automatically change Vulkan present method changed to native when transparency support is enabled.
