+++ platform/windows/SCsub
@@ -27,6 +27,7 @@ common_win = [
     "wgl_detect_version.cpp",
     "rendering_context_driver_vulkan_windows.cpp",
     "drop_target_windows.cpp",
+    "nvapi_profile.cpp",
 ]
 
 if env.msvc:
