+++ platform/windows/gl_manager_windows_native.cpp
@@ -35,7 +35,7 @@
 #include "core/config/project_settings.h"
 #include "core/version.h"
 
-#include "thirdparty/misc/nvapi_minimal.h"
+#include "nvapi_profile.h"
 
 #include <dwmapi.h>
 #include <cstdio>
@@ -66,220 +66,6 @@ static String format_error_message(DWORD id) {
 	LocalFree(messageBuffer);
 
 	return msg;
-}
-
-const int OGL_THREAD_CONTROL_ID = 0x20C1221E;
-const int OGL_THREAD_CONTROL_DISABLE = 0x00000002;
-const int OGL_THREAD_CONTROL_ENABLE = 0x00000001;
-const int VRR_MODE_ID = 0x1194F158;
-const int VRR_MODE_FULLSCREEN_ONLY = 0x1;
-
-typedef int(__cdecl *NvAPI_Initialize_t)();
-typedef int(__cdecl *NvAPI_Unload_t)();
-typedef int(__cdecl *NvAPI_GetErrorMessage_t)(unsigned int, NvAPI_ShortString);
-typedef int(__cdecl *NvAPI_DRS_CreateSession_t)(NvDRSSessionHandle *);
-typedef int(__cdecl *NvAPI_DRS_DestroySession_t)(NvDRSSessionHandle);
-typedef int(__cdecl *NvAPI_DRS_LoadSettings_t)(NvDRSSessionHandle);
-typedef int(__cdecl *NvAPI_DRS_CreateProfile_t)(NvDRSSessionHandle, NVDRS_PROFILE *, NvDRSProfileHandle *);
-typedef int(__cdecl *NvAPI_DRS_CreateApplication_t)(NvDRSSessionHandle, NvDRSProfileHandle, NVDRS_APPLICATION *);
-typedef int(__cdecl *NvAPI_DRS_SaveSettings_t)(NvDRSSessionHandle);
-typedef int(__cdecl *NvAPI_DRS_SetSetting_t)(NvDRSSessionHandle, NvDRSProfileHandle, NVDRS_SETTING *);
-typedef int(__cdecl *NvAPI_DRS_FindProfileByName_t)(NvDRSSessionHandle, NvAPI_UnicodeString, NvDRSProfileHandle *);
-typedef int(__cdecl *NvAPI_DRS_GetApplicationInfo_t)(NvDRSSessionHandle, NvDRSProfileHandle, NvAPI_UnicodeString, NVDRS_APPLICATION *);
-typedef int(__cdecl *NvAPI_DRS_DeleteProfile_t)(NvDRSSessionHandle, NvDRSProfileHandle);
-NvAPI_GetErrorMessage_t NvAPI_GetErrorMessage__;
-
-static bool nvapi_err_check(const char *msg, int status) {
-	if (status != 0) {
-		if (OS::get_singleton()->is_stdout_verbose()) {
-			NvAPI_ShortString err_desc = { 0 };
-			NvAPI_GetErrorMessage__(status, err_desc);
-			print_verbose(vformat("%s: %s(code %d)", msg, err_desc, status));
-		}
-		return false;
-	}
-	return true;
-}
-
-// On windows we have to customize the NVIDIA application profile:
-// * disable threaded optimization when using NVIDIA cards to avoid stuttering, see
-//   https://stackoverflow.com/questions/36959508/nvidia-graphics-driver-causing-noticeable-frame-stuttering/37632948
-//   https://github.com/Ryujinx/Ryujinx/blob/master/src/Ryujinx.Common/GraphicsDriver/NVThreadedOptimization.cs
-// * disable G-SYNC in windowed mode, as it results in unstable editor refresh rates
-void GLManagerNative_Windows::_nvapi_setup_profile() {
-	HMODULE nvapi = nullptr;
-#ifdef _WIN64
-	nvapi = LoadLibraryA("nvapi64.dll");
-#else
-	nvapi = LoadLibraryA("nvapi.dll");
-#endif
-
-	if (nvapi == nullptr) {
-		return;
-	}
-
-	void *(__cdecl * NvAPI_QueryInterface)(unsigned int interface_id) = nullptr;
-
-	NvAPI_QueryInterface = (void *(__cdecl *)(unsigned int))(void *)GetProcAddress(nvapi, "nvapi_QueryInterface");
-
-	if (NvAPI_QueryInterface == nullptr) {
-		print_verbose("Error getting NVAPI NvAPI_QueryInterface");
-		return;
-	}
-
-	// Setup NVAPI function pointers
-	NvAPI_Initialize_t NvAPI_Initialize = (NvAPI_Initialize_t)NvAPI_QueryInterface(0x0150E828);
-	NvAPI_GetErrorMessage__ = (NvAPI_GetErrorMessage_t)NvAPI_QueryInterface(0x6C2D048C);
-	NvAPI_DRS_CreateSession_t NvAPI_DRS_CreateSession = (NvAPI_DRS_CreateSession_t)NvAPI_QueryInterface(0x0694D52E);
-	NvAPI_DRS_DestroySession_t NvAPI_DRS_DestroySession = (NvAPI_DRS_DestroySession_t)NvAPI_QueryInterface(0xDAD9CFF8);
-	NvAPI_Unload_t NvAPI_Unload = (NvAPI_Unload_t)NvAPI_QueryInterface(0xD22BDD7E);
-	NvAPI_DRS_LoadSettings_t NvAPI_DRS_LoadSettings = (NvAPI_DRS_LoadSettings_t)NvAPI_QueryInterface(0x375DBD6B);
-	NvAPI_DRS_CreateProfile_t NvAPI_DRS_CreateProfile = (NvAPI_DRS_CreateProfile_t)NvAPI_QueryInterface(0xCC176068);
-	NvAPI_DRS_CreateApplication_t NvAPI_DRS_CreateApplication = (NvAPI_DRS_CreateApplication_t)NvAPI_QueryInterface(0x4347A9DE);
-	NvAPI_DRS_SaveSettings_t NvAPI_DRS_SaveSettings = (NvAPI_DRS_SaveSettings_t)NvAPI_QueryInterface(0xFCBC7E14);
-	NvAPI_DRS_SetSetting_t NvAPI_DRS_SetSetting = (NvAPI_DRS_SetSetting_t)NvAPI_QueryInterface(0x577DD202);
-	NvAPI_DRS_FindProfileByName_t NvAPI_DRS_FindProfileByName = (NvAPI_DRS_FindProfileByName_t)NvAPI_QueryInterface(0x7E4A9A0B);
-	NvAPI_DRS_GetApplicationInfo_t NvAPI_DRS_GetApplicationInfo = (NvAPI_DRS_GetApplicationInfo_t)NvAPI_QueryInterface(0xED1F8C69);
-	NvAPI_DRS_DeleteProfile_t NvAPI_DRS_DeleteProfile = (NvAPI_DRS_DeleteProfile_t)NvAPI_QueryInterface(0x17093206);
-
-	if (!nvapi_err_check("NVAPI: Init failed", NvAPI_Initialize())) {
-		return;
-	}
-
-	print_verbose("NVAPI: Init OK!");
-
-	NvDRSSessionHandle session_handle;
-
-	if (NvAPI_DRS_CreateSession == nullptr) {
-		return;
-	}
-
-	if (!nvapi_err_check("NVAPI: Error creating DRS session", NvAPI_DRS_CreateSession(&session_handle))) {
-		NvAPI_Unload();
-		return;
-	}
-
-	if (!nvapi_err_check("NVAPI: Error loading DRS settings", NvAPI_DRS_LoadSettings(session_handle))) {
-		NvAPI_DRS_DestroySession(session_handle);
-		NvAPI_Unload();
-		return;
-	}
-
-	String app_executable_name = OS::get_singleton()->get_executable_path().get_file();
-	String app_profile_name = GLOBAL_GET("application/config/name");
-	// We need a name anyways, so let's use the engine name if an application name is not available
-	// (this is used mostly by the Project Manager)
-	if (app_profile_name.is_empty()) {
-		app_profile_name = GODOT_VERSION_NAME;
-	}
-	String old_profile_name = app_profile_name + " Nvidia Profile";
-	Char16String app_profile_name_u16 = app_profile_name.utf16();
-	Char16String old_profile_name_u16 = old_profile_name.utf16();
-	Char16String app_executable_name_u16 = app_executable_name.utf16();
-
-	// A previous error in app creation logic could result in invalid profiles,
-	// clean these if they exist before proceeding.
-	NvDRSProfileHandle old_profile_handle;
-
-	int old_status = NvAPI_DRS_FindProfileByName(session_handle, (NvU16 *)(old_profile_name_u16.ptrw()), &old_profile_handle);
-
-	if (old_status == 0) {
-		print_verbose("NVAPI: Deleting old profile...");
-
-		if (!nvapi_err_check("NVAPI: Error deleting old profile", NvAPI_DRS_DeleteProfile(session_handle, old_profile_handle))) {
-			NvAPI_DRS_DestroySession(session_handle);
-			NvAPI_Unload();
-			return;
-		}
-
-		if (!nvapi_err_check("NVAPI: Error deleting old profile", NvAPI_DRS_SaveSettings(session_handle))) {
-			NvAPI_DRS_DestroySession(session_handle);
-			NvAPI_Unload();
-			return;
-		}
-	}
-
-	NvDRSProfileHandle profile_handle = nullptr;
-
-	int profile_status = NvAPI_DRS_FindProfileByName(session_handle, (NvU16 *)(app_profile_name_u16.ptrw()), &profile_handle);
-
-	if (profile_status != 0) {
-		print_verbose("NVAPI: Profile not found, creating...");
-
-		NVDRS_PROFILE profile_info;
-		profile_info.version = NVDRS_PROFILE_VER;
-		profile_info.isPredefined = 0;
-		memcpy(profile_info.profileName, app_profile_name_u16.get_data(), sizeof(char16_t) * app_profile_name_u16.size());
-
-		if (!nvapi_err_check("NVAPI: Error creating profile", NvAPI_DRS_CreateProfile(session_handle, &profile_info, &profile_handle))) {
-			NvAPI_DRS_DestroySession(session_handle);
-			NvAPI_Unload();
-			return;
-		}
-	}
-
-	NVDRS_APPLICATION_V4 app;
-	app.version = NVDRS_APPLICATION_VER_V4;
-
-	int app_status = NvAPI_DRS_GetApplicationInfo(session_handle, profile_handle, (NvU16 *)(app_executable_name_u16.ptrw()), &app);
-
-	if (app_status != 0) {
-		print_verbose("NVAPI: Application not found in profile, creating...");
-
-		app.isPredefined = 0;
-		memcpy(app.appName, app_executable_name_u16.get_data(), sizeof(char16_t) * app_executable_name_u16.size());
-		memcpy(app.launcher, L"", sizeof(wchar_t));
-		memcpy(app.fileInFolder, L"", sizeof(wchar_t));
-
-		if (!nvapi_err_check("NVAPI: Error creating application", NvAPI_DRS_CreateApplication(session_handle, profile_handle, &app))) {
-			NvAPI_DRS_DestroySession(session_handle);
-			NvAPI_Unload();
-			return;
-		}
-	}
-
-	NVDRS_SETTING ogl_thread_control_setting = {};
-	ogl_thread_control_setting.version = NVDRS_SETTING_VER;
-	ogl_thread_control_setting.settingId = OGL_THREAD_CONTROL_ID;
-	ogl_thread_control_setting.settingType = NVDRS_DWORD_TYPE;
-	int thread_control_val = OGL_THREAD_CONTROL_DISABLE;
-	if (!GLOBAL_GET("rendering/gl_compatibility/nvidia_disable_threaded_optimization")) {
-		thread_control_val = OGL_THREAD_CONTROL_ENABLE;
-	}
-	ogl_thread_control_setting.u32CurrentValue = thread_control_val;
-
-	if (!nvapi_err_check("NVAPI: Error calling NvAPI_DRS_SetSetting", NvAPI_DRS_SetSetting(session_handle, profile_handle, &ogl_thread_control_setting))) {
-		NvAPI_DRS_DestroySession(session_handle);
-		NvAPI_Unload();
-		return;
-	}
-
-	NVDRS_SETTING vrr_mode_setting = {};
-	vrr_mode_setting.version = NVDRS_SETTING_VER;
-	vrr_mode_setting.settingId = VRR_MODE_ID;
-	vrr_mode_setting.settingType = NVDRS_DWORD_TYPE;
-	vrr_mode_setting.u32CurrentValue = VRR_MODE_FULLSCREEN_ONLY;
-
-	if (!nvapi_err_check("NVAPI: Error calling NvAPI_DRS_SetSetting", NvAPI_DRS_SetSetting(session_handle, profile_handle, &vrr_mode_setting))) {
-		NvAPI_DRS_DestroySession(session_handle);
-		NvAPI_Unload();
-		return;
-	}
-
-	if (!nvapi_err_check("NVAPI: Error saving settings", NvAPI_DRS_SaveSettings(session_handle))) {
-		NvAPI_DRS_DestroySession(session_handle);
-		NvAPI_Unload();
-		return;
-	}
-
-	if (thread_control_val == OGL_THREAD_CONTROL_DISABLE) {
-		print_verbose("NVAPI: Disabled OpenGL threaded optimization successfully");
-	} else {
-		print_verbose("NVAPI: Enabled OpenGL threaded optimization successfully");
-	}
-	print_verbose("NVAPI: Disabled G-SYNC for windowed mode successfully");
-
-	NvAPI_DRS_DestroySession(session_handle);
 }
 
 int GLManagerNative_Windows::_find_or_create_display(GLWindow &win) {
@@ -503,7 +289,21 @@ void GLManagerNative_Windows::swap_buffers() {
 }
 
 Error GLManagerNative_Windows::initialize() {
-	_nvapi_setup_profile();
+	Dictionary props;
+	int thread_control_val = OGL_THREAD_CONTROL_DISABLE;
+	if (!GLOBAL_GET("rendering/gl_compatibility/nvidia_disable_threaded_optimization")) {
+		thread_control_val = OGL_THREAD_CONTROL_ENABLE;
+	}
+	props[OGL_THREAD_CONTROL_ID] = thread_control_val;
+	props[VRR_MODE_ID] = VRR_MODE_FULLSCREEN_ONLY;
+	if (nvapi_setup_profile(props)) {
+		if (thread_control_val == OGL_THREAD_CONTROL_DISABLE) {
+			print_verbose("NVAPI: Disabled OpenGL threaded optimization successfully");
+		} else {
+			print_verbose("NVAPI: Enabled OpenGL threaded optimization successfully");
+		}
+		print_verbose("NVAPI: Disabled G-SYNC for windowed mode successfully");
+	}
 	return OK;
 }
